Rearrange the imports and things so that one can write `use rayon::par_iter::*` and get just the traits with methods you want to call. I went ahead and made the submodules public just to avoid quandries about unnameable paths, but it's not clear that we want to expose all those contents, just because most users will not care.
